pyproject.toml
[tool.poetry]
name = "church-backend"
version = "0.1.0"
description = ""
authors = ["Ihor Palii <darkfree97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
scrapy = "^2.11.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
adrf = "^0.1.7"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
